+++ concordium-node/src/grpc2.rs
@@ -705,6 +705,8 @@ struct ServiceConfig {
     #[serde(default)]
     get_block_certificates: bool,
     #[serde(default)]
+    get_baker_earliest_win_time: bool,
+    #[serde(default)]
     get_first_block_epoch: bool,
     #[serde(default)]
     get_winning_bakers_epoch: bool,
@@ -765,6 +767,7 @@ impl ServiceConfig {
             get_block_items: true,
             get_bakers_reward_period: true,
             get_block_certificates: true,
+            get_baker_earliest_win_time: true,
             get_first_block_epoch: true,
             get_winning_bakers_epoch: true,
         }
@@ -1858,6 +1861,20 @@ pub mod server {
             let hash = self.consensus.get_bakers_reward_period_v2(request.get_ref(), sender)?;
             let mut response = tonic::Response::new(receiver);
             add_hash(&mut response, hash)?;
+            Ok(response)
+        }
+
+        async fn get_baker_earliest_win_time(
+            &self,
+            request: tonic::Request<crate::grpc2::types::BakerId>,
+        ) -> Result<tonic::Response<Vec<u8>>, tonic::Status> {
+            if !self.service_config.get_baker_earliest_win_time {
+                return Err(tonic::Status::unimplemented(
+                    "`GetBakerEarliestWinTime` is not enabled.",
+                ));
+            }
+            let response = self.consensus.get_baker_earliest_win_time_v2(request.get_ref())?;
+            let response = tonic::Response::new(response);
             Ok(response)
         }